+++ src/ts/Ads/Managers/AssetManager.ts
@@ -16,6 +16,7 @@ import { Diagnostics } from 'Core/Utilities/Diagnostics';
 import { PerformanceCampaign } from 'Performance/Models/PerformanceCampaign';
 import { XPromoCampaign } from 'XPromo/Models/XPromoCampaign';
 import { CreativeBlocking, BlockingReason } from 'Core/Utilities/CreativeBlocking';
+import { PromoCampaign } from 'Promo/Models/PromoCampaign';
 
 enum CacheType {
     REQUIRED,
@@ -87,6 +88,13 @@ export class AssetManager {
     }
 
     public setup(campaign: Campaign): Promise<Campaign> {
+        if (campaign instanceof PromoCampaign) {
+            // only run this logic for promo campaigns
+            if (this.shouldDisableCacheForPromo(navigator.userAgent)) {
+                return Promise.resolve(campaign);
+            }
+        }
+
         if (this._cacheMode === CacheMode.DISABLED) {
             return Promise.resolve(campaign);
         }
@@ -390,5 +398,19 @@ export class AssetManager {
         CreativeBlocking.report(campaign.getCreativeId(), seatId, campaign.getId(), BlockingReason.FILE_TOO_LARGE, {
             fileSize: Math.floor(totalSize / (1024 * 1024))
         });
+    }
+
+    /**
+     * Promos are currently broken in the latest Android Webview 77 Release.
+     * TODO: Once this issue is resolved, then this workaround should be removed
+     */
+    private shouldDisableCacheForPromo(userAgent: unknown) {
+        const userAgentString = (userAgent && typeof userAgent === 'string') ? userAgent : '';
+        const regexFields = userAgentString.match(/Chrom(e|ium)\/([0-9]+)/);
+        if (regexFields && regexFields.length >= 3) {
+            const majorVersion = parseInt(regexFields[2], 10);
+            return majorVersion >= 77;
+        }
+        return false;
     }
 }