+++ src/ts/Core/Models/ABGroup.ts
@@ -68,7 +68,6 @@ export const FakeZyngaFilteredABTest = new ZyngaFilteredABTest(16, 17);
 export const OpenMeasurementTest = new DisabledABTest();
 export const IframeEndcardTest = new DisabledABTest();
 export const HtmlEndcardTest = new DisabledABTest();
-export const SliderEndCardExperiment = new ZyngaFilteredABTest(7, 8);
 export const ZyngaWWFLoadTest = new ABTest(13, 14);
 export const MediationLoadExperiment = new ABTest(15, 16);
 export const ConsentUXTest = new ABTest(18, 19);